requirements.txt
einops
wandb
scikit-learn
python-hostlist
tqdm
requests
pyyaml
timm == 0.4.12
mmcv==1.3.8
mmsegmentation==0.14.1